% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_filter.R
\name{REDCap_filter}
\alias{REDCap_filter}
\title{Filter REDCap dataset}
\usage{
REDCap_filter(
  data,
  data_dictionary,
  id_variable,
  forms = NULL,
  variables = NULL
)
}
\arguments{
\item{data}{Dataset to filter}

\item{data_dictionary}{Data dictionary}

\item{id_variable}{Name of the identifier variable}

\item{forms}{Forms to include; defaults to all}

\item{variables}{Variables to include; defaults to all}
}
\value{
A filtered dataframe
}
\description{
Filter REDCap dataset
}
\examples{
REDCap_filter(data = REDCap_data, data_dictionary = REDCap_dd, id_variable = "record_id")
REDCap_filter(data = REDCap_data, data_dictionary = REDCap_dd, id_variable = "record_id", forms = c("form1", "form2", "..."), variables = c("var1", "var2", "..."))
}
