% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_process_branching_logic.R
\name{REDCap_logic_parser}
\alias{REDCap_logic_parser}
\title{Translate REDCap logic to R}
\usage{
REDCap_logic_parser(data_dictionary, column_name, missing_data_codes = NULL)
}
\arguments{
\item{data_dictionary}{REDCap data dictionary}

\item{column_name}{Name of the column containing containing the REDCap logic in the data dictionary}

\item{missing_data_codes}{Codes used to specify missing data; defaults to c(-777, -888, -999, -7777, -8888, -9999, NA)}
}
\description{
Translate REDCap logic to R
}
\examples{
REDCap_logic_parser(data_dictionary = REDCap_data_dictionary, column_name = "branching_logic")
REDCap_logic_parser(data_dictionary = dd_export, column_name = "branching logic", missing_data_codes = c("-9999", "-999"))
}
