% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_map_labels.R
\name{REDCap_map_labels}
\alias{REDCap_map_labels}
\title{Map metadata labels to data values}
\usage{
REDCap_map_labels(data, dictionary, variables = "All")
}
\arguments{
\item{data}{The REDCap data to map}

\item{dictionary}{The REDCap data dictionary}

\item{variables}{Variable(s) to map; defaults to all}
}
\value{
A dataset with mapped values
}
\description{
Allows to map caterogical values of a REDCap dataset to their respective data dictionary labels. Only applicable to radio and dropdown variables.
}
\examples{
REDCap_map_names(data = REDCap_data, dictionary = REDCap_dd, variables = "All")
}
