% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_map_names.R
\name{REDCap_map_names}
\alias{REDCap_map_names}
\title{Map variable label to variable name}
\usage{
REDCap_map_names(data, dictionary, variables = "All")
}
\arguments{
\item{data}{The REDCap data to map}

\item{dictionary}{The REDCap data dictionary}

\item{variables}{Variable(s) to map; defaults to all}
}
\value{
A dataset with mapped labels
}
\description{
Map variable label to variable name
}
\examples{
REDCap_map_names(data = REDCap_data, dictionary = REDCap_dd, variables = "All")
}
