% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_missing_value_report.R
\name{REDCap_missing_value_report}
\alias{REDCap_missing_value_report}
\title{This function generates a missing variables report}
\usage{
REDCap_missing_value_report(
  ...,
  data = NULL,
  dictionary = NULL,
  event_form = NULL,
  InstrumentSkip = NULL,
  event_description = NULL,
  id_variable = "record_id",
  start_of_tx_variable = NULL,
  include_complete_table = FALSE
)
}
\arguments{
\item{...}{List containing the REDCap data, dictionary and event form mapping (if required).
This should be the output of the \code{REDCap_import} function.}

\item{data}{Data frame containing the data read from REDCap.}

\item{dictionary}{Data frame containing the REDCap dictionary.}

\item{event_form}{Data frame mapping each form to each event.}

\item{InstrumentSkip}{Data frame specifying forms to skip under certain conditions.}

\item{event_description}{Data frame describing each event.}

\item{id_variable}{Character string naming the participant ID variable.}

\item{start_of_tx_variable}{Character string naming the variable containing the start of treatment date.}

\item{include_complete_table}{Logical, whether to return the complete table.}
}
\value{
A data frame or a list of data frames.
}
\description{
This function generates a missing variables report
}
