% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_parse_validation.R
\name{REDCap_parse_validation}
\alias{REDCap_parse_validation}
\title{Validate REDcap to R translation}
\usage{
REDCap_parse_validation(data, logic, column_name)
}
\arguments{
\item{data}{REDCap data on which you want to validate the parsing}

\item{logic}{Dataset or string containing the R logic}

\item{column_name}{Name of the column containing the logic if "logic" is a dataset, else, leave blank}
}
\description{
This function allows you to validate the output of REDCap_logic_parser in the context of you REDCap data
}
