% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_missing_value_report.R
\name{REDCap_missing_value_report}
\alias{REDCap_missing_value_report}
\title{This function generates a missing variables report}
\usage{
REDCap_missing_value_report(
  ...,
  data = NULL,
  dictionary = NULL,
  event_form = NULL,
  InstrumentSkip = NULL,
  event_description = NULL,
  id_variable = "record_id",
  start_of_tx_variable
)
}
\arguments{
\item{...}{List containing the REDCap data, dictionary and event form mapping (if required). This should be the output of the \code{REDCap_import} function.}

\item{data}{Data frame containing the data read from REDCap. If a list (...) is specified, this argument is not required.}

\item{dictionary}{Data frame containing the REDCap dictionary. If a list (...) is specified, this argument is not required.}

\item{event_form}{Data frame containing the mapping of each form with each event. If a list (...) is specified, this argument is not required.}

\item{InstrumentSkip}{Data frame containing the mapping of each form with each event. If a list (...) is specified, this argument is not required.}

\item{event_description}{Data frame containing the description of each event. If a list (...) is specified, this argument is not required.}

\item{id_variable}{Character vector of lenght 1 containing the name of the ID variable (identifier).}

\item{start_of_tx_variable}{Character vector of lenght 1 containing the name of the variable containing the start of treatment date (identifier).}
}
\description{
This function generates a missing variables report
}
